feat: add article information section to main article view
Added a new component, that displays various bits of article and journal metadata in the main article view. I've had to make some less than ideal compromises to get things working quickly, namely some changes to the JATS importer that now mean that the 'history' elements are no longer parsed. This will need to be fixed in the future, but the importer and metadata models need some serious re-working to generalise date/pub-date elements.